Remove "unreachable" as notification
As per http://docs.icinga.org/latest/en/networkreachability.html the
"unreachable" state is determined by the network logic of parent/child
relationships.

We do not want to alert on "unreachable" states if we're aware that when
a parent is down it will affect monitoring to other things (i.e. if a
VPN gateway is down then all the hosts behind it are effectively "down",
at least according to the management nodes).